fix more doc links
