+++ src/events.rs
@@ -6,10 +6,10 @@ These can be used with [Dispatcher::subscribe][Subscribe] to get a [Stream] of [
 
 See the [table]
 
-[Subscribe]: ../runner/struct.Dispatcher.html#method.subscribe
-[Stream]: https://docs.rs/futures/0.3.1/futures/stream/trait.Stream.html
+[Subscribe]: ../struct.Dispatcher.html#method.subscribe
+[Stream]: https://docs.rs/tokio/0.2/tokio/stream/trait.Stream.html
 [Messages]: ../messages/index.html
-[table]: ../runner/struct.Dispatcher.html#a-table-of-mappings
+[table]: ../struct.Dispatcher.html#a-table-of-mappings
 */
 use super::messages;
 use crate::runner::Event;
@@ -225,7 +225,7 @@ impl<'t> Event<'t> for All {
 }
 
 // TODO generate this with a macro
-use crate::runner::Dispatcher;
+use crate::Dispatcher;
 pub(crate) fn build_event_map(dispatcher: Dispatcher) -> Dispatcher {
     // TODO this acquires the lock N times
     // if we expose the inner map
