+++ src/messages/mod.rs
@@ -5,9 +5,9 @@ This can be obtained by [subscribing] to an [Event] on a [Dispatcher]
 
 Or by using [Parse] on an [Message]
 
-[subscribing]: ../runner/struct.Dispatcher.html#method.subscribe
+[subscribing]: ../struct.Dispatcher.html#method.subscribe
 [Event]: ../events/index.html
-[Dispatcher]: ../runner/struct.Dispatcher.html
+[Dispatcher]: ../struct.Dispatcher.html
 [Parse]: ../trait.Parse.html
 [Message]: ../decode/struct.Message.html
 */
