+++ src/runner/control.rs
@@ -1,4 +1,4 @@
-use super::*;
+use {super::*, crate::*};
 
 #[derive(Clone)]
 /// A control type for writing messages to the client, or stopping it.
